Fix for an invalid 'to map threshold' value being passed in (cherry picked from commit 2814eee6fac2b1f5d7c17dbbc1a49273ff50a463)
Conflicts:

	kernel/src/main/java/org/neo4j/kernel/impl/transaction/ReadOnlyTxManager.java
	kernel/src/main/java/org/neo4j/kernel/impl/transaction/TxManager.java

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/core/RelationshipTypeHolder.java
@@ -35,7 +35,7 @@ import org.neo4j.kernel.impl.util.ArrayMap;
 public class RelationshipTypeHolder
 {
     private ArrayMap<String,Integer> relTypes = 
-        new ArrayMap<String,Integer>( 5, true, true );
+        new ArrayMap<String,Integer>( (byte)5, true, true );
     private Map<Integer, RelationshipTypeImpl> relTranslation = new ConcurrentHashMap<Integer, RelationshipTypeImpl>();
 
     private final TransactionManager transactionManager;
