+++ community/kernel/src/main/java/org/neo4j/kernel/impl/transaction/TxManager.java
@@ -142,7 +142,7 @@ public class TxManager extends AbstractTransactionManager
     public void init( XaDataSourceManager xaDsManagerToUse )
     {
         this.xaDsManager = xaDsManagerToUse;
-        txThreadMap = new ArrayMap<Thread,TransactionImpl>( 5, true, true );
+        txThreadMap = new ArrayMap<Thread,TransactionImpl>( (byte)5, true, true );
         logSwitcherFileName = txLogDir + separator + "active_tx_log";
         txLog1FileName = "tm_tx_log.1";
         txLog2FileName = "tm_tx_log.2";
